+++ src/main/kotlin/in/kkkev/jjidea/actions/change/describeAction.kt
@@ -1,19 +1,19 @@
 package `in`.kkkev.jjidea.actions.change
 
-import com.intellij.icons.AllIcons
 import com.intellij.openapi.project.Project
 import `in`.kkkev.jjidea.actions.nullAndDumbAwareAction
 import `in`.kkkev.jjidea.actions.requestDescription
 import `in`.kkkev.jjidea.jj.Description
 import `in`.kkkev.jjidea.jj.LogEntry
 import `in`.kkkev.jjidea.jj.invalidate
+import `in`.kkkev.jjidea.ui.common.JujutsuIcons
 
 /**
  * Describe action.
  * Opens a dialog to edit the description of a revision.
  */
 fun describeAction(project: Project, logEntry: LogEntry?) =
-    nullAndDumbAwareAction(logEntry, "log.action.describe", AllIcons.Actions.Edit) {
+    nullAndDumbAwareAction(logEntry, "log.action.describe", JujutsuIcons.Describe) {
         val jujutsuRoot = target.repo
         val commandExecutor = jujutsuRoot.commandExecutor